[insight] remove duplicated code
