+++ src/PJM/AppBundle/Datatables/BaseDatatable.php
@@ -4,12 +4,32 @@ namespace PJM\AppBundle\Datatables;
 
 use Sg\DatatablesBundle\Datatable\View\AbstractDatatableView;
 use Sg\DatatablesBundle\Datatable\View\Style;
+use PJM\AppBundle\Twig\IntranetExtension;
 
 /**
  * Class BaseDatatable.
  */
 abstract class BaseDatatable extends AbstractDatatableView
 {
+    protected $intranetExt;
+    protected $ajaxUrl;
+    protected $admin;
+
+    public function setIntranetExt(IntranetExtension $intranetExt)
+    {
+        $this->intranetExt = $intranetExt;
+    }
+
+    public function setAjaxUrl($ajaxUrl)
+    {
+        $this->ajaxUrl = $ajaxUrl;
+    }
+
+    public function setAdmin($admin)
+    {
+        $this->admin = $admin;
+    }
+
     public function buildDatatable()
     {
         $this->options->setOptions(array(
